requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pillow>=9.0.0
gunicorn>=21.0.0
